refactor: lane is uint8 in Go (column stays smallint)
The lane only ever holds small non-negative class ids (0 = fast, 1 = slow);
uint8 states that in the type. LaneFast/LaneSlow consts, ClassifyLane,
the CheckJob model field, and the claim/release plumbing all follow. No
schema change — Postgres has no unsigned types, so the column remains
smallint and bun scans it into uint8 unchanged.

+++ server/internal/checkworker/checkjobsvc/service.go
@@ -76,7 +76,7 @@ type Service interface {
 		costEWMAMs float64,
 		delayEWMAMs float64,
 		effectiveScheduledAt time.Time,
-		lane int16,
+		lane uint8,
 	) error
 }
 
@@ -318,7 +318,7 @@ func (s *serviceImpl) selectAvailableJobs(
 	tx bun.Tx,
 	jobs *[]*models.CheckJob,
 	region *string,
-	lane int16,
+	lane uint8,
 	limit int,
 	maxAhead time.Duration,
 	now time.Time,
@@ -487,7 +487,7 @@ func (s *serviceImpl) ReleaseLeaseWithSchedulingState(
 	costEWMAMs float64,
 	delayEWMAMs float64,
 	effectiveScheduledAt time.Time,
-	lane int16,
+	lane uint8,
 ) error {
 	update := s.db.NewUpdate().
 		Model((*models.CheckJob)(nil)).
